core-to-plc: move some modules to more logical locations
