+++ core-to-plc/src/Language/Plutus/CoreToPLC/Compiler/Names.hs
@@ -5,11 +5,12 @@
 -- | Functions for compiling GHC names into Plutus Core names.
 module Language.Plutus.CoreToPLC.Compiler.Names where
 
+import           PlutusPrelude                            (strToBs)
+
 import           Language.Plutus.CoreToPLC.Compiler.Kind
 import {-# SOURCE #-} Language.Plutus.CoreToPLC.Compiler.Type
 import           Language.Plutus.CoreToPLC.Compiler.Types
 import           Language.Plutus.CoreToPLC.PLCTypes
-import           Language.Plutus.CoreToPLC.Utils
 
 import qualified GhcPlugins                               as GHC
