+++ core-to-plc/src/Language/Plutus/CoreToPLC/Compiler/Types.hs
@@ -5,24 +5,24 @@
 
 module Language.Plutus.CoreToPLC.Compiler.Types where
 
-import           Language.Plutus.CoreToPLC.Error
+import           Language.Plutus.CoreToPLC.Compiler.Error
 import           Language.Plutus.CoreToPLC.PIRTypes
 import           Language.Plutus.CoreToPLC.PLCTypes
 
 import           Language.PlutusCore.Quote
 
-import qualified GhcPlugins                         as GHC
+import qualified GhcPlugins                               as GHC
 
 import           Control.Monad.Except
 import           Control.Monad.Reader
 import           Control.Monad.State
 
-import qualified Data.List.NonEmpty                 as NE
-import qualified Data.Map                           as Map
-import qualified Data.Set                           as Set
+import qualified Data.List.NonEmpty                       as NE
+import qualified Data.Map                                 as Map
+import qualified Data.Set                                 as Set
 import           Lens.Micro
 
-import qualified Language.Haskell.TH.Syntax         as TH
+import qualified Language.Haskell.TH.Syntax               as TH
 
 type BuiltinNameInfo = Map.Map TH.Name GHC.TyThing
